add generic email attachment finder and remove duplicated boilerplate
